actualizar

+++ deb/etc/init.d/impresora.sh
@@ -1,17 +1,19 @@
 #!/bin/bash
 
-case "$1" in
-start)
-echo "Iniciando Servicio..."
-./opt/impresora/contador.sh
-;;
-stop)
-echo "Deteniendo Servicio..."
-pkill -f contador.sh
-;;
-*)
-echo "Modeo de empleo: /etc/init.d/impresora.sh {start|stop}"
-exit 1
-;;
-esac
-exit 0
+if [[ $# -eq 0 ]]
+then
+	echo "Modeo de empleo: /etc/init.d/impresora.sh {start|stop}"
+elif [[ $1 = "start" ]]
+then
+	echo "Iniciando Servicio..."
+	cd /opt/impresora
+	./contador.sh
+else
+	echo "Deteniendo Servicio..."
+	pkill -f contador.sh
+fi
+
+
+
+
+
